split everything into components fixed lint errors
