+++ src/components/LayoutComponents/Footer.js
@@ -1,0 +1,14 @@
+import React from "react";
+import {Layout} from "antd";
+const {Footer} = Layout
+
+export const FooterComponent = ({text}) => {
+  return(
+    <Footer
+      className="text-center"
+    >
+      {text}
+    </Footer>
+  )
+}
+
