+++ src/index.jsx
@@ -1,19 +1,15 @@
-/* eslint-disable */
 import ReactDOM from "react-dom/client";
 import React from "react";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import MainPage from "./components/MainPage";
-import "./index.css"
+import "./index.css";
 import {LayoutComponent} from "./components/Layout";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
 	<>
-	<App /> 
-		<LayoutComponent>
-			<MainPage />
-		</LayoutComponent>
+		<App /> 
+		<LayoutComponent />
 	</>
 );
